tools/requirements.txt
tqdm
pyserial
